+++ main-service/src/main/java/ru/practicum/dto/event/EventShortDto.java
@@ -24,7 +24,7 @@ public class EventShortDto {
     private String annotation;
     @NotNull
     private CategoryResponseDto category;
-    private Long confirmedRequests;
+    private Integer confirmedRequests;
     @NotNull
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime eventDate;
@@ -33,5 +33,5 @@ public class EventShortDto {
     @NotNull
     @Size(min = 3, max = 120)
     private String title;
-    private Long views;
+    private Integer views;
 }
